requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Pillow>=9.0